Refactor StrongHttpsClient to move all SOCKS logic into ClientConnOperator
